requirements.txt

Flask==1.0.2
click==6.7
Flask-SQLAlchemy==2.3.2
Jinja2==2.10
MarkupSafe==1.0
SQLAlchemy==1.2.12
Werkzeug==0.14.1
itsdangerous==0.24
PyMySQL
python-dotenv
